Remove pytumblr dependency
Pytumblr has never been properly updated to accomodate Python 3, so
installing it is unnecesarily fiddly. Competing library Tumblpy doesn't
offer a way to specify a single post ID when requesting posts.

Now, no Tumblr library is used, and a direct HTTP request to the Tumblr
API is created with the requests library.