Seperate command line interface and connect parts
The command line interfaces are now handled through a single file
tumblr2jekyll.py, which treats the other two files as libraries. This
allows their functionality to be connected together.

This allows a script to be written which goes through every post on a
blog, traces its posts back to when they were reblogged by a particular
user, and saves the versions posted by that user.

The specific use case for this is that I post my work on my main blog,
and reblog particular posts to a side blog. I want to use the sideblog
as a guide to which posts from the main blog need to be retrieved.

The command line output is still very messy due to debugging, and needs
cleanup.